+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added keybindings to effect the throttle of the rocket
 - Add level indicator to navball
 - Add velocity markers to navball
+- Add frame around navball
 
 # v0.4.0 - 2019-08-24
 - Clamp camera altitude to [1 mm, 1 Tm]